+++ tech/Reflection/PropertyDef.h
@@ -28,6 +28,8 @@ THE SOFTWARE.
 #include "Reflection/Property.h"
 #include "Reflection/Object.h"
 
+#include <cstdint>
+
 namespace Teardrop
 {
 	namespace Reflection
@@ -118,14 +120,14 @@ namespace Teardrop
 				if (index >= 0)
 				{
 					CollectionPropertyType<_T>* pColl = 
-						(CollectionPropertyType<_T>*)((unsigned long)pObj + m_offset);
+                        (CollectionPropertyType<_T>*)((uintptr_t)pObj + m_offset);
 					if (pColl->getCount() <= size_t(index))
 						pColl->resize(index+1);
 					pColl->setAtFromString(index, pVal);
 				}
 				else
 				{
-					PropertyDefImpl<_T>* pProp = (PropertyDefImpl<_T>*)((unsigned long)pObj + m_offset);
+                    PropertyDefImpl<_T>* pProp = (PropertyDefImpl<_T>*)((uintptr_t)pObj + m_offset);
 					pProp->fromString(pVal);
 				}
 
@@ -134,25 +136,25 @@ namespace Teardrop
 
 			void setData(Object* pObj, const void* val) const
 			{
-				PropertyDefImpl<_T>* pProp = (PropertyDefImpl<_T>*)((unsigned long)pObj + m_offset);
+                PropertyDefImpl<_T>* pProp = (PropertyDefImpl<_T>*)((uintptr_t)pObj + m_offset);
 				pProp->set(*((_T*)val));
 				pObj->notifyPropertyChanged(this);
 			}
 
 			void getData(const Object* pObj, const void* val) const
 			{
-				PropertyDefImpl<_T>* pProp = (PropertyDefImpl<_T>*)((unsigned long)pObj + m_offset);
+                PropertyDefImpl<_T>* pProp = (PropertyDefImpl<_T>*)((uintptr_t)pObj + m_offset);
 				*((_T*)val) = pProp->get();
 			}
 
 			const void* getDataPointer(const Object* pObj) const
 			{
-				return (const void*)((unsigned long)pObj + m_offset);
+                return (const void*)((uintptr_t)pObj + m_offset);
 			}
 
 			void getDataAsString(const Object* pObj, String& sVal) const
 			{
-				PropertyDefImpl<_T>* pProp = (PropertyDefImpl<_T>*)((unsigned long)pObj + m_offset);
+                PropertyDefImpl<_T>* pProp = (PropertyDefImpl<_T>*)((uintptr_t)pObj + m_offset);
 				pProp->toString(sVal);
 			}
 
@@ -160,14 +162,14 @@ namespace Teardrop
 			{
 				if (m_bCollection)
 				{
-					CollectionPropertyType<_T>* pProp = (CollectionPropertyType<_T>*)((unsigned long)pSrc + m_offset);
-					CollectionPropertyType<_T>* pOtherProp = (CollectionPropertyType<_T>*)((unsigned long)pDest + m_offset);
+                    CollectionPropertyType<_T>* pProp = (CollectionPropertyType<_T>*)((uintptr_t)pSrc + m_offset);
+                    CollectionPropertyType<_T>* pOtherProp = (CollectionPropertyType<_T>*)((uintptr_t)pDest + m_offset);
 					*pOtherProp = *pProp;
 				}
 				else
 				{
-					PropertyDefImpl<_T>* pProp = (PropertyDefImpl<_T>*)((unsigned long)pSrc + m_offset);
-					PropertyDefImpl<_T>* pOtherProp = (PropertyDefImpl<_T>*)((unsigned long)pDest + m_offset);
+                    PropertyDefImpl<_T>* pProp = (PropertyDefImpl<_T>*)((uintptr_t)pSrc + m_offset);
+                    PropertyDefImpl<_T>* pOtherProp = (PropertyDefImpl<_T>*)((uintptr_t)pDest + m_offset);
 					*pOtherProp = *pProp;
 				}
 			}
@@ -211,7 +213,7 @@ namespace Teardrop
 
 			const void* getDataPointer(const Object* pObj) const
 			{
-				return (const void*)((unsigned long)pObj + m_offset);
+                return (const void*)((uintptr_t)pObj + m_offset);
 			}
 
 			void getDataAsString(const Object* /*pObj*/, String& /*sVal*/) const
@@ -250,7 +252,7 @@ namespace Teardrop
 
 			void setData(Object* pObj, const void* v) const
 			{
-				PointerPropertyDefImpl<_T>* pProp = (PointerPropertyDefImpl<_T>*)((unsigned long)pObj + m_offset);
+                PointerPropertyDefImpl<_T>* pProp = (PointerPropertyDefImpl<_T>*)((uintptr_t)pObj + m_offset);
 				_T* val = (_T*)v;
 				(*pProp) = val;
 				pObj->notifyPropertyChanged(this);
@@ -258,19 +260,19 @@ namespace Teardrop
 
 			void getData(const Object* pObj, const void* v) const
 			{
-				PointerPropertyDefImpl<_T>& prop = *((PointerPropertyDefImpl<_T>*)((unsigned long)pObj + m_offset));
+                PointerPropertyDefImpl<_T>& prop = *((PointerPropertyDefImpl<_T>*)((uintptr_t)pObj + m_offset));
 				_T** val = (_T**)v;
 				*val = prop;
 			}
 
 			const void* getDataPointer(const Object* pObj) const
 			{
-				return (const void*)((unsigned long)pObj + m_offset);
+                return (const void*)((uintptr_t)pObj + m_offset);
 			}
 
 			void getDataAsString(const Object* pObj, String& sVal) const
 			{
-				PointerPropertyDefImpl<_T>& prop = *((PointerPropertyDefImpl<_T>*)((unsigned long)pObj + m_offset));
+                PointerPropertyDefImpl<_T>& prop = *((PointerPropertyDefImpl<_T>*)((uintptr_t)pObj + m_offset));
 				Reflection::Object* obj = (Reflection::Object*)(_T*)prop;
 				if (obj)
 					obj->getObjectId().toString(sVal);
@@ -278,8 +280,8 @@ namespace Teardrop
 
 			void copyTo(Object* pDest, const Object* pSrc) const
 			{
-				PointerPropertyDefImpl<_T>* pProp = (PointerPropertyDefImpl<_T>*)((unsigned long)pSrc + m_offset);
-				PointerPropertyDefImpl<_T>* pOtherProp = (PointerPropertyDefImpl<_T>*)((unsigned long)pDest + m_offset);
+                PointerPropertyDefImpl<_T>* pProp = (PointerPropertyDefImpl<_T>*)((uintptr_t)pSrc + m_offset);
+                PointerPropertyDefImpl<_T>* pOtherProp = (PointerPropertyDefImpl<_T>*)((uintptr_t)pDest + m_offset);
 				*pOtherProp = *pProp;
 			}
 
@@ -312,7 +314,7 @@ namespace Teardrop
 					const EnumValue* enumVal = enumDef->values();
 					while (enumVal) {
 						if (String(enumVal->id()) == str) {
-							EnumPropertyDefImpl<T>* pProp = (EnumPropertyDefImpl<T>*)((unsigned long)pObj + m_offset);
+                            EnumPropertyDefImpl<T>* pProp = (EnumPropertyDefImpl<T>*)((uintptr_t)pObj + m_offset);
 							*pProp = (T)enumVal->value();
 							pObj->notifyPropertyChanged(this);
 						}
@@ -324,7 +326,7 @@ namespace Teardrop
 
 			void setData(Object* pObj, const void* v) const
 			{
-				EnumPropertyDefImpl<T>* pProp = (EnumPropertyDefImpl<T>*)((unsigned long)pObj + m_offset);
+                EnumPropertyDefImpl<T>* pProp = (EnumPropertyDefImpl<T>*)((uintptr_t)pObj + m_offset);
 				int val = *((int*)v);
 				*pProp = (T)val;
 				pObj->notifyPropertyChanged(this);
@@ -332,19 +334,19 @@ namespace Teardrop
 
 			void getData(const Object* pObj, const void* v) const
 			{
-				EnumPropertyDefImpl<T>& prop = *((EnumPropertyDefImpl<T>*)((unsigned long)pObj + m_offset));
+                EnumPropertyDefImpl<T>& prop = *((EnumPropertyDefImpl<T>*)((uintptr_t)pObj + m_offset));
 				int* val = (int*)v;
 				*val = int(prop);
 			}
 
 			const void* getDataPointer(const Object* pObj) const
 			{
-				return (const void*)((unsigned long)pObj + m_offset);
+                return (const void*)((uintptr_t)pObj + m_offset);
 			}
 
 			void getDataAsString(const Object* pObj, String& sVal) const
 			{
-				EnumPropertyDefImpl<T>& prop = *((EnumPropertyDefImpl<T>*)((unsigned long)pObj + m_offset));
+                EnumPropertyDefImpl<T>& prop = *((EnumPropertyDefImpl<T>*)((uintptr_t)pObj + m_offset));
 				T val = prop;
 
 				const ClassDef* classDef = pObj->getDerivedClassDef();
@@ -363,8 +365,8 @@ namespace Teardrop
 
 			void copyTo(Object* pDest, const Object* pSrc) const
 			{
-				EnumPropertyDefImpl<T>* pProp = (EnumPropertyDefImpl<T>*)((unsigned long)pSrc + m_offset);
-				EnumPropertyDefImpl<T>* pOtherProp = (EnumPropertyDefImpl<T>*)((unsigned long)pDest + m_offset);
+                EnumPropertyDefImpl<T>* pProp = (EnumPropertyDefImpl<T>*)((uintptr_t)pSrc + m_offset);
+                EnumPropertyDefImpl<T>* pOtherProp = (EnumPropertyDefImpl<T>*)((uintptr_t)pDest + m_offset);
 				*pOtherProp = *pProp;
 			}
